transfer models:MasterClass, Performance, Reading and all related components from library app to afisha app
