+++ apps/afisha/admin/masterclass.py
@@ -1,0 +1,16 @@
+from django.contrib import admin
+
+from apps.afisha.admin.performance import TeamMemberInline
+from apps.afisha.models.masterclass import MasterClass
+
+
+@admin.register(MasterClass)
+class MasterClassAdmin(admin.ModelAdmin):
+    list_display = ("name",)
+    exclude = ("events",)
+    search_fields = (
+        "project",
+        "play__name",
+        "name",
+    )
+    inlines = (TeamMemberInline,)
