+++ apps/afisha/factories/__init__.py
@@ -1,0 +1,11 @@
+from .events import EventFactory
+from .masterclass import MasterClassFactory
+from .performance import PerformanceFactory
+from .reading import ReadingFactory
+
+__all__ = (
+    EventFactory,
+    MasterClassFactory,
+    PerformanceFactory,
+    ReadingFactory,
+)
