+++ apps/afisha/serializers/__init__.py
@@ -1,0 +1,17 @@
+from .masterclass import EventMasterClassSerializer
+from .performance import (
+    EventPerformanceSerializer,
+    PerformanceMediaReviewSerializer,
+    PerformanceReviewSerializer,
+    PerformanceSerializer,
+)
+from .reading import EventReadingSerializer
+
+__all__ = (
+    EventMasterClassSerializer,
+    EventPerformanceSerializer,
+    EventReadingSerializer,
+    PerformanceSerializer,
+    PerformanceReviewSerializer,
+    PerformanceMediaReviewSerializer,
+)
