+++ apps/afisha/serializers/events.py
@@ -1,9 +1,11 @@
 from drf_spectacular.utils import PolymorphicProxySerializer, extend_schema_field
 from rest_framework import serializers
 
-from apps.afisha.models import Event
-from apps.library.models import MasterClass, Performance, Reading
-from apps.library.serializers import EventMasterClassSerializer, EventPerformanceSerializer, EventReadingSerializer
+from apps.afisha.models.events import Event
+from apps.afisha.models.masterclass import MasterClass
+from apps.afisha.models.performance import Performance
+from apps.afisha.models.reading import Reading
+from apps.afisha.serializers import EventMasterClassSerializer, EventPerformanceSerializer, EventReadingSerializer
 
 AFISHA_EVENTS_SERIALIZER_PAIRS = {
     MasterClass: EventMasterClassSerializer,
