+++ apps/afisha/serializers/reading.py
@@ -1,17 +1,17 @@
 from rest_framework import serializers
 
-from apps.library.models import MasterClass
+from apps.afisha.models.reading import Reading
 from apps.library.serializers.role import RoleSerializer
 
 
-class EventMasterClassSerializer(serializers.ModelSerializer):
-    """Master-class serializer for afisha page."""
+class EventReadingSerializer(serializers.ModelSerializer):
+    """Reading serializer for afisha page."""
 
     team = RoleSerializer(source="event_team", many=True)
     project_title = serializers.SlugRelatedField(slug_field="title", read_only=True, source="project")
 
     class Meta:
-        model = MasterClass
+        model = Reading
         fields = (
             "id",
             "name",
