+++ apps/afisha/views/__init__.py
@@ -1,0 +1,16 @@
+from .events import AfishaEventListAPIView, AfishaInfoAPIView
+from .performance import (
+    PerformanceMediaReviewViewSet,
+    PerformancePreviewDetailAPI,
+    PerformanceReviewViewSet,
+    PerformanceViewSet,
+)
+
+__all__ = (
+    PerformanceMediaReviewViewSet,
+    PerformancePreviewDetailAPI,
+    PerformanceReviewViewSet,
+    PerformanceViewSet,
+    AfishaEventListAPIView,
+    AfishaInfoAPIView,
+)
