+++ apps/afisha/views/performance.py
@@ -4,13 +4,10 @@ from rest_framework import mixins, viewsets
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from apps.afisha.models.performance import Performance
+from apps.afisha.serializers import PerformanceMediaReviewSerializer, PerformanceReviewSerializer, PerformanceSerializer
 from apps.articles import selectors
-from apps.library.models import Performance, PerformanceMediaReview, PerformanceReview
-from apps.library.serializers import (
-    PerformanceMediaReviewSerializer,
-    PerformanceReviewSerializer,
-    PerformanceSerializer,
-)
+from apps.library.models import PerformanceMediaReview, PerformanceReview
 
 
 class PerformanceViewSet(mixins.RetrieveModelMixin, viewsets.GenericViewSet):
