+++ apps/articles/models/__init__.py
@@ -1,6 +1,6 @@
-from .blog_items import BlogItem, BlogItemContent, BlogPerson
-from .news_items import NewsItem, NewsItemContent
-from .projects import Project, ProjectContent
+from apps.articles.models.blog_items import BlogItem, BlogItemContent, BlogPerson
+from apps.articles.models.news_items import NewsItem, NewsItemContent
+from apps.articles.models.projects import Project, ProjectContent
 
 __all__ = (
     BlogItem,
