+++ apps/content_pages/models/__init__.py
@@ -1,5 +1,13 @@
-from .content_items import AbstractItemWithTitle, ContentUnitRichText, Link, Preamble, Quote, Text, Title
-from .contents import AbstractContent, AbstractContentPage
+from apps.content_pages.models.content_items import (
+    AbstractItemWithTitle,
+    ContentUnitRichText,
+    Link,
+    Preamble,
+    Quote,
+    Text,
+    Title,
+)
+from apps.content_pages.models.contents import AbstractContent, AbstractContentPage
 
 # Prevent isort to rearrange imports and prevent circular imports.
 # isort: split
