+++ apps/content_pages/serializers/content_blocks.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 
-from apps.afisha.models import Event
+from apps.afisha.models.events import Event
+from apps.afisha.serializers.performance import EventPerformanceSerializer
 from apps.content_pages.models import EventsBlock, ImagesBlock, PersonsBlock, PlaysBlock, VideosBlock
 from apps.content_pages.serializers import (
     ExtendedPersonSerializer,
@@ -8,7 +9,6 @@ from apps.content_pages.serializers import (
     OrderedPlaySerializer,
     OrderedVideoSerializer,
 )
-from apps.library.serializers.performance import EventPerformanceSerializer
 
 
 class EventInBlockSerializer(serializers.ModelSerializer):
