+++ apps/core/management/commands/filldb.py
@@ -2,7 +2,8 @@ from typing import Any, Optional
 
 from django.core.management.base import BaseCommand, CommandError
 
-from apps.afisha.factories import EventFactory
+from apps.afisha.factories import MasterClassFactory, PerformanceFactory, ReadingFactory
+from apps.afisha.factories.events import EventFactory
 from apps.core.factories import PersonFactory
 from apps.core.models import Setting
 from apps.feedback.factories import ParticipationApplicationFestivalFactory
@@ -17,15 +18,7 @@ from apps.info.factories import (
     VolunteerFactory,
 )
 from apps.info.models import FestivalTeamMember
-from apps.library.factories import (
-    AuthorFactory,
-    MasterClassFactory,
-    OtherPlayFactory,
-    PerformanceFactory,
-    PlayFactory,
-    ProgramTypeFactory,
-    ReadingFactory,
-)
+from apps.library.factories import AuthorFactory, OtherPlayFactory, PlayFactory, ProgramTypeFactory
 from apps.main.factories import BannerFactory as MainBannerFactory
 from apps.users.factories import AdminUserFactory, EditorUserFactory, JournalistUserFactory, ObserverUserFactory
