+++ apps/library/admin/__init__.py
@@ -1,8 +1,4 @@
-from django.contrib import admin
-from django.contrib.sites.models import Site
-
 from .author import AuthorAdmin
-from .events import MasterClassAdmin, PerformanceAdmin, ReadingAdmin
 from .performancemediareview import PerformanceMediaReviewAdmin
 from .performancereview import PerformanceReviewAdmin
 from .play import PlayAdmin
@@ -10,13 +6,8 @@ from .programtype import ProgramTypeAdmin
 
 __all__ = (
     AuthorAdmin,
-    MasterClassAdmin,
-    PerformanceAdmin,
     PerformanceMediaReviewAdmin,
     PerformanceReviewAdmin,
     PlayAdmin,
     ProgramTypeAdmin,
-    ReadingAdmin,
 )
-
-admin.site.unregister(Site)
