+++ apps/library/factories/__init__.py
@@ -1,6 +1,3 @@
 from .author import AuthorFactory
-from .masterclass import MasterClassFactory
-from .performance import PerformanceFactory
 from .play import OtherPlayFactory, PlayFactory, ProgramTypeFactory
-from .reading import ReadingFactory
 from .team_member import TeamMemberFactory
