+++ apps/library/models/__init__.py
@@ -1,23 +1,18 @@
 from .author import Achievement, Author, AuthorPlay, OtherLink, SocialNetworkLink
-from .master_class import MasterClass
-from .performance import Performance, PerformanceImage, PerformanceMediaReview, PerformanceReview
+from .performancemediareview import PerformanceImage, PerformanceMediaReview, PerformanceReview
 from .play import Play, ProgramType
-from .reading import Reading
 from .team_member import TeamMember
 
 __all__ = (
     Achievement,
     Author,
     AuthorPlay,
-    MasterClass,
     OtherLink,
-    Performance,
     PerformanceImage,
     PerformanceMediaReview,
     PerformanceReview,
     Play,
     ProgramType,
-    Reading,
     SocialNetworkLink,
     TeamMember,
 )
