+++ apps/library/serializers/__init__.py
@@ -1,14 +1,6 @@
 from .author import AuthorLettersSerializer, AuthorListSerializer, AuthorRetrieveSerializer, AuthorSearchSerializer
-from .masterclass import EventMasterClassSerializer
-from .performance import (
-    EventPerformanceSerializer,
-    PerformanceMediaReviewSerializer,
-    PerformanceReviewSerializer,
-    PerformanceSerializer,
-)
 from .play import AuthorForPlaySerializer, AuthorPlaySerializer, PlaySerializer
 from .playfilters import PlayFiltersSerializer
-from .reading import EventReadingSerializer
 from .role import RoleSerializer
 
 __all__ = (
@@ -18,13 +10,7 @@ __all__ = (
     AuthorSearchSerializer,
     AuthorLettersSerializer,
     AuthorForPlaySerializer,
-    EventMasterClassSerializer,
-    EventPerformanceSerializer,
-    EventReadingSerializer,
     PlaySerializer,
     PlayFiltersSerializer,
-    PerformanceSerializer,
-    PerformanceReviewSerializer,
-    PerformanceMediaReviewSerializer,
     RoleSerializer,
 )
