+++ apps/library/urls.py
@@ -1,24 +1,16 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
+from apps.afisha.views import PerformanceMediaReviewViewSet, PerformanceReviewViewSet
 from apps.library.views import (
     AuthorLettersAPIView,
     AuthorsReadViewSet,
-    PerformanceMediaReviewViewSet,
-    PerformancePreviewDetailAPI,
-    PerformanceReviewViewSet,
-    PerformanceViewSet,
     PlayFiltersAPIView,
     PlayViewSet,
     SearchResultViewSet,
 )
 
 router = DefaultRouter()
-router.register(
-    "performances",
-    PerformanceViewSet,
-    basename="performance",
-)
 router.register(
     r"performances/(?P<performance_id>\d+)/media-reviews",
     PerformanceMediaReviewViewSet,
@@ -45,17 +37,8 @@ router.register(
     AuthorsReadViewSet,
     basename="authors",
 )
-performance_item_urls = [
-    path("", include(router.urls)),
-    path(
-        route="performances/<int:id>/preview/",
-        view=PerformancePreviewDetailAPI.as_view(),
-        name="performance-detail-preview",
-    ),
-]
 
 library_urls = [
-    path("library/", include(performance_item_urls)),
     path(
         "library/author_letters/",
         AuthorLettersAPIView.as_view(),
