+++ apps/library/views/__init__.py
@@ -1,10 +1,4 @@
 from .author import AuthorLettersAPIView, AuthorsReadViewSet
-from .performance import (
-    PerformanceMediaReviewViewSet,
-    PerformancePreviewDetailAPI,
-    PerformanceReviewViewSet,
-    PerformanceViewSet,
-)
 from .play import PlayViewSet
 from .playfilters import PlayFiltersAPIView
 from .searchresult import SearchResultViewSet
@@ -12,11 +6,7 @@ from .searchresult import SearchResultViewSet
 __all__ = (
     AuthorsReadViewSet,
     PlayFiltersAPIView,
-    PerformanceViewSet,
     PlayViewSet,
     SearchResultViewSet,
-    PerformancePreviewDetailAPI,
-    PerformanceReviewViewSet,
-    PerformanceMediaReviewViewSet,
     AuthorLettersAPIView,
 )
